+++ client/src/components/App.js
@@ -7,7 +7,7 @@ import Header from './header/Header'
 import LoginPage from './pages/LoginPage'
 import Dashboard from './pages/Dashboard'
 import Statistics from './pages/Statistics'
-import Logs from './pages/Logs'
+import Logs from './pages/LogsPage'
 import InsertUser from './pages/InsertUserPage'
 import ModifyUser from './pages/ModifyUserPage'
 import DeleteUser from './pages/DeleteUserPage'
