+++ client/src/components/pages/LogsPage.js
